requirements.txt
azure-cognitiveservices-speech==1.35.0
